pyproject.toml
[project]
name = "binance-trader"
version = "0.1.0"
description = "AXS market monitor and trader for Binance"
requires-python = ">=3.12,<3.15"
dependencies = [
    "binance-sdk-spot>=7.0.0",
    "binance-sdk-derivatives-trading-usds-futures>=7.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp-socks>=0.9.0",
    "PySocks>=1.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.0.0",
    "mcp[cli] (>=1.0.0)",
    "anthropic (>=0.40.0)",
    "asyncpg>=0.29.0",
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "trader"}, {include = "db"}]
